all: move to google.cloud namespace
To ensure publishing and importing as

    from google.cloud.sqlcommenter import
