+++ python/sqlcommenter-python/google/cloud/sqlcommenter/psycopg2/extension.py
@@ -16,10 +16,9 @@
 
 import psycopg2
 import psycopg2.extensions
-
-from sqlcommenter import generate_sql_comment
-from sqlcommenter.flask import get_flask_info
-from sqlcommenter.opencensus import get_opencensus_values
+from google.cloud.sqlcommenter import generate_sql_comment
+from google.cloud.sqlcommenter.flask import get_flask_info
+from google.cloud.sqlcommenter.opencensus import get_opencensus_values
 
 
 # This integration extends psycopg2.extensions.cursor
