+++ python/sqlcommenter-python/tests/generic/test_generate_sql_comment.py
@@ -16,7 +16,7 @@
 
 from unittest import TestCase
 
-from sqlcommenter import generate_sql_comment
+from google.cloud.sqlcommenter import generate_sql_comment
 
 
 class GenerateSqlCommentTests(TestCase):
