+++ python/sqlcommenter-python/tests/sqlalchemy/__init__.py
@@ -13,4 +13,3 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
-
