+++ python/sqlcommenter-python/tests/sqlalchemy/tests.py
@@ -17,8 +17,7 @@
 from unittest import TestCase
 
 import sqlalchemy
-
-from sqlcommenter.sqlalchemy.executor import BeforeExecuteFactory
+from google.cloud.sqlcommenter.sqlalchemy.executor import BeforeExecuteFactory
 
 from ..compat import mock
 from ..opencensus_mock import mock_opencensus_tracer
@@ -78,27 +77,27 @@ class FlaskTests(SQLAlchemyTestCase):
         'route': '/',
     }
 
-    @mock.patch('sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_all_data(self, get_info):
         self.assertSQL(
             "SELECT 1; /*controller='c',framework='flask',route='/'*/",
         )
 
-    @mock.patch('sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_framework_disabled(self, get_info):
         self.assertSQL(
             "SELECT 1; /*controller='c',route='/'*/",
             with_framework=False,
         )
 
-    @mock.patch('sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_controller_disabled(self, get_info):
         self.assertSQL(
             "SELECT 1; /*framework='flask',route='/'*/",
             with_controller=False,
         )
 
-    @mock.patch('sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
+    @mock.patch('google.cloud.sqlcommenter.sqlalchemy.executor.get_flask_info', return_value=flask_info)
     def test_route_disabled(self, get_info):
         self.assertSQL(
             "SELECT 1; /*controller='c',framework='flask'*/",
